+++ tradingmate/ui/gtk/ConfirmDialog.py
@@ -1,3 +1,5 @@
+# flake8: noqa: E402 # Required to allow use of gi.require_version
+
 import gi
 
 gi.require_version("Gtk", "3.0")