+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Added Makefile to perform development and deployment actions
 - Added formatting with black and isort
+- Linting with flake8
 
 ## Removed
 - Removed `setup.py` with full usage of `pyproject.toml`